services/auth-service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pyotp
pydantic[email]>=2